backend/requirements.txt
fastapi
python-dotenv
uvicorn
langchain
langchain_community
langchain_groq
openai
langchain_openai

pwdlib
jwt
passlib[argon2]
sqlalchemy
python-jose[cryptography]
authlib
itsdangerous